Move the `testing-server` module under the `@bayou` namespace.
