+++ local-modules/@bayou/testing-client/client-tests
@@ -4,5 +4,5 @@
 
 The contents of this file don't matter. Its name triggers special-case logic
 in the Webpack config defined in `client-bundle`, which in turn calls on
-`testing-server.ClientTestsLoader.load()` to build a dynamic module that
+`@bayou/testing-server.ClientTestsLoader.load()` to build a dynamic module that
 references all of the client test code.
