+++ local-modules/@bayou/testing-server/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "testing-server",
+  "name": "@bayou/testing-server",
   "version": "1.0.0",
 
   "dependencies": {
