+++ local-modules/client-bundle/package.json
@@ -7,7 +7,7 @@
     "@bayou/deps-util-server": "local",
     "@bayou/env-server": "local",
     "@bayou/see-all": "local",
-    "testing-server": "local",
+    "@bayou/testing-server": "local",
     "@bayou/util-common": "local"
   }
 }
