+++ server/package.json
@@ -25,6 +25,6 @@
     "@bayou/promise-util": "local",
     "@bayou/see-all": "local",
     "@bayou/see-all-server": "local",
-    "testing-server": "local"
+    "@bayou/testing-server": "local"
   }
 }
